// src/types/ConfigAuditReportReportChecks.ts
/**
 * 
 * 
 *
 * The version of the OpenAPI document: 
 * Contact Email: 
 * License: 
 *
 * NOTE: This file is auto generated by crdtotypes (https://github.com/yaacov/crdtoapi/).
 * https://github.com/yaacov/crdtoapi/README.crdtotypes
 */

import { ConfigAuditReportReportChecksScope } from './ConfigAuditReportReportChecksScope';

/**
 * Check provides the result of conducting a single audit step.
 *
 * @export
 */
export interface ConfigAuditReportReportChecks {
  /** category
   *
   * @required {false}
   */
  category?: string;
  /** checkID
   *
   * @required {true}
   */
  checkID: string;
  /** description
   *
   * @required {false}
   */
  description?: string;
  /** messages
   *
   * @required {false}
   */
  messages?: string[];
  /** remediation
   * Remediation provides description or links to external resources to remediate failing check.
   *
   * @required {false}
   */
  remediation?: string;
  /** scope
   * Scope indicates the section of config that was audited.
   *
   * @required {false}
   */
  scope?: ConfigAuditReportReportChecksScope;
  /** severity
   * Severity level of a vulnerability or a configuration audit check.
   *
   * @required {true}
   */
  severity: string;
  /** success
   *
   * @required {true}
   */
  success: boolean;
  /** title
   *
   * @required {false}
   */
  title?: string;
}